feat: unsave warning

+++ src/modules/workshop/UI/Panel/BottomBar/index.tsx
@@ -159,8 +159,8 @@ export default function BottomBar() {
 
   const handleGetData = async () => {
     const data = (await getListModularByWallet({
-      ownerAddress: account?.address,
-      // ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
+      // ownerAddress: account?.address,
+      ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
       page: 1,
       limit: 20,
     })) as any
@@ -170,9 +170,10 @@ export default function BottomBar() {
   }
 
   const handleClickCreateNewProject = () => {
-    // if (blocksState.length > 2) {
-    //   setShowUnsaveModal(true)
-    // }
+    if (isAllowSave) {
+      setShowUnsaveModal(true)
+      return
+    }
     createProject()
     deleteAlls()
     handleGetData()
@@ -268,7 +269,7 @@ export default function BottomBar() {
           </button>
         </div>
       </div>
-      {/* <UnsaveWarningModal show={showUnsaveModal} setIsOpen={setShowUnsaveModal} /> */}
+      <UnsaveWarningModal show={showUnsaveModal} setIsOpen={setShowUnsaveModal} />
     </>
   )
 }
